refactor : DTO 변환 로직을 DTO 내부로 변경
