+++ src/main/java/chess/view/dto/PieceResponse.java
@@ -1,4 +1,4 @@
-package chess.dto;
+package chess.view.dto;
 
 import chess.domain.piece.Color;
 import chess.domain.piece.Type;
